requirements.txt
Django==2.0
djangorestframework==3.8.2
drf-yasg==1.9.1
Pillow=5.1.0